Update README and CHANGELOGS with new version info
The README file was updated with the information about the new version for both Ocarina of Time and Majora's Mask Trackers. The versioning scheme was changed and the features planned for the next versions were described. In the CHANGELOGS, details about the new features implemented in the new versions were added.